examples/requirements.txt
fastapi == 0.101.0
uvicorn[standard] == 0.23.2
Jinja2 == 3.1.2
cassandra-driver==3.29.0